+++ posthog-server/src/test/java/com/posthog/server/internal/PostHogFeatureFlagsTest.kt
@@ -221,7 +221,7 @@ internal class PostHogFeatureFlagsTest {
             )
 
         assertNull(result)
-        assertTrue(logger.containsLog("Loading remote feature flags failed"))
+        assertTrue(logger.containsLog("Loading remote feature flags API error"))
         mockServer.shutdown()
     }