+++ posthog-server/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - feat: Add ETag support for local evaluation polling to reduce bandwidth when flags haven't changed ([#350](https://github.com/PostHog/posthog-android/pull/350))
+- feat: `$feature_flag_called` events now report `$feature_flag_error` property ([#355](https://github.com/PostHog/posthog-android/pull/355))
 
 ## 2.1.0 - 2025-12-05